perf(fri): set default FRI arity to 4
Change GoldilocksCubicProofOptions default fri_log_arity from 1 to 2
(arity-4). This halves the number of FRI Merkle trees committed during
the proof, reducing FRI Merkle hashing work by ~50%.

All 132 tests pass including arity-specific roundtrip tests.

+++ crypto/stark/src/verifier.rs
@@ -744,7 +744,7 @@ pub trait IsStarkVerifier<
             // Each successive sub-fold halves `half`, so the product `leaf_index * half` correctly
             // tracks the pair position in the correspondingly smaller domain.
             let mut inv_twiddles: Vec<FieldElement<Field>> = Vec::with_capacity(half);
-            let half_log_bits = (current_domain_size / 2) as u64;
+            let half_log_bits = current_domain_size / 2;
             for j in 0..half {
                 let global_pair_idx = leaf_index * half + j;
                 let natural_idx = reverse_index(global_pair_idx, half_log_bits);
